Add partly functional `np()` (improved)
Added fetching of album art to `np()` the 'now playing' command.
It's a bit of a hit-or-miss though, most likely because it
is difficult to know which part of the youtube title is
the track title and which is the artist, leading to an incorrect
track search, which sometimes fails.
Needs to be fixed in later versions.